// VERDEVIA-MOBILE-BACKEND/src/modules/users/enums/user.enums.ts
export enum UserGender {
  MASCULINO = 'Masculino',
  FEMININO = 'Feminino',
  OUTRO = 'Outro',
  NAO_INFORMAR = 'Prefiro não dizer',
}

export enum UserEthnicity {
  BRANCA = 'Branca',
  PRETA = 'Preta',
  PARDA = 'Parda',
  AMARELA = 'Amarela',
  INDIGENA = 'Indígena',
}

export enum UserRole {
  SUPER_ADMIN = 'super_admin',
  ADMIN = 'admin',
  USER = 'user',
  MODERATOR = 'moderator',
  CONTRACTOR = 'contractor',
  SUPER_CONTRACTOR = 'super_contractor',
}